feat(neofetch): Logic for determining image renderer
This helps to work out whether we can generate pretty images or whether
we should convert images to ASCII.

As this is dependant on the features of our terminal (ie. Tmux has
issues rendering images within Kitty) we have to make sure that we're
converting images to ASCII whenever we're in Tmux.

Signed-off-by: Yomi Colledge <baphled@boodah.net>