feat(fastfetch): Replace neofetch
We were having issues with neofetch causing tmuxinator to not be able to
setup our sessions correctly.

fastfetch, as it's name implies, is much quicker and does not suffer
from the same issues.

We've had to make some tweaks to our setup, to get it to work reasonably
well, but that is a small sacrifice considering it doesn't not gobble up
commands whilst its trying to load.

We do have the downside of not being able to dynamically generate ASCII
art using our images, but this may change at some point, dependant on
our our free time on the progress of the project as a whole.

Signed-off-by: Yomi Colledge <baphled@boodah.net>