feat(touchegg): Add basic functionality for touchpad
We want to be able to customise our touchpad, so that we're able to use
it as an extension to our keyboard.

We already have quite a bit of flexibility with our keyboard, but
sometimes we still want to be able to navigate with some form of mouse.
Using a touchpad allows us to have a device that we can extend in a
similar way to our custom keyboard, so that we can tailor it to our own
needs.

Signed-off-by: Yomi Colledge <baphled@boodah.net>