feat(alacritty): Add configuration and custom theme
We're not sure whether we'll want to use this as our replacement
terminal, but we already prefer it to kitty. So adding the basics to
determine whether we want to make it our terminal of choice.

Added catppuccin-mocha, so that it falls in like with the rest of our
colour schemes.

Signed-off-by: Yomi Colledge <baphled@boodah.net>