feat(sway): Add config to handle idling
When we idle, we also need to make sure that we're able to turn our
monitor off. Otherwise we'll end up with screen burn.

This is an attempt to get us closer to being able to handle this.
Although, we still need to confirm that this works as expected. So it
may need further tweaking.

We will need to come back and define a proper idle screen, and
potentially look at having some sort of screen saver. Which would wrap
this functionality up nicely.

Signed-off-by: Yomi Colledge <baphled@boodah.net>