feat(waybar): Attempt to get waybar to display media information
We want to be able to see what kind of media we are playing in the
background. It should be possible for waybar to show us this
information, but it will take some playing around with.

This branch is exclusively setup to explore this, and play with the
settings to get the look and feel just right.

Signed-off-by: Yomi Colledge <baphled@boodah.net>