feat(sway): Improve upon auto loading applications
This makes sure that we start all of our basic applications, and ensures
that are display on the correct workspaces.

We may want to tweak this further, by having different `app_id`s to
determine which workspace we'd like other parts of our setup to go. But
this serves as a decent starting to point to allow us to get things
going and not be hindered by having to manually start these commonly
used applications up, anytime that we start a new session.

Signed-off-by: Yomi Colledge <baphled@boodah.net>