fix(sway): Move mod key to Super, so Alt does not hide our tmux options
This means that we're going to have to get used to how we navigate our
sway workspaces, but it also means that we're able to navigate through
our tmux sessions also.

We've been using tmux _much_ longer, so the change should be more of a
benefit that a hindrance.

It also means that we'll be able to navigate around our browser more
efficiently as our browsers also use 'Alt' for jumping between tabs.

Signed-off-by: Yomi Colledge <baphled@boodah.net>