feat(gpg): Add custom gpg script
Used to work out whether we have access to $DISPLAY or not.

If $DISPLAY is enabled we use the system gpg to sign commits, otherwise
we use pinetry-mode loopback.

pinetry-mode loopback allows us to fake inquires back to the caller, so
that we're able to submit a passphrase via the terminal.

Signed-off-by: Yomi Colledge <baphled@boodah.net>