feat(hyprland): Integrate into workflow
We like sway, but the issues with screen flickering is a bit _too_ much.
So we've been experiementing with hypland. We've found that it's has
better support for nvidia and it's quite nice on the eyes too.

This commit works towards integrating hyprland into our workflow and
capturing the same functionality we get from sway into hyprland. Whilst
making sure that we're not missing anything major in our migration away
from sway.

Signed-off-by: Yomi Colledge <baphled@boodah.net>