fix(fzf): Improve usability
As we've made a few changes to our shell, we need to give fzf some love
and make sure that it's working as we expect.

This includes properly integrating with tmux, and having a more
consistent setup regardless of the command we're using.

This allows us to think less about what we want to do, and relying on
conventions we're used to. This includes improving how we do tab
completions, as well as how we handle CTRL-T, CTRL-R, and ALT-C.

Signed-off-by: Yomi Colledge <baphled@boodah.net>