fix(zsh): Improve load time
Here we make improves in regards to how long it takes for our shell to
start up.

We introduced the ability for 1password to load our secrets
automatically, but this caused our shell to take a very long time to
load.

Along with this, we also had nvm not lazy loaded. Which is also another
module that caused considerable slow down.

So here, we've refactored how we handle initialising op, to make sure
that we're able to load the shell, without having to rely upon op being
run.

It would be nice to make sure that we have a visual cue for this, but
we'll deal with that separately, sot hhat this commit only focuses on
performance improvements.

Signed-off-by: Yomi Colledge <baphled@boodah.net>